requirements.txt
aioredis==1.3.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
Deprecated==1.2.18
dnspython==2.7.0
email-validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.13
fastapi-cli==0.0.7
h11==0.16.0
hiredis==3.2.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jinja2==3.1.6
limits==4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==24.2
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.2
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.7
shellingham==1.5.4
slowapi==0.1.9
sniffio==1.3.1
starlette==0.46.2
typer==0.16.0
typing-extensions==4.14.0
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.2
resend~=2.10.0